requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
cryptography==43.0.0
